requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
networkx>=3.1
matplotlib>=3.7.2
numpy>=1.24.3
pillow>=10.0.0
python-dateutil>=2.8.2
pandas>=2.0.3